+++ examples/native-sdk-ios-client/NativeWeb3App/AppDelegate.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import CoinbaseWalletSDK
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
