+++ packages/wallet-native-sdk/ios/Message/Message.swift
@@ -12,21 +12,8 @@ import CryptoKit
 public typealias Message<C> = BaseMessage<C> where C: UnencryptedContent
 
 @available(iOS 13.0, *)
-public protocol UnencryptedContent: BaseContent {
+protocol UnencryptedContent: BaseContent {
     associatedtype Encrypted: EncryptedContent where Encrypted.Unencrypted == Self
     
     init(decrypt encrypted: Encrypted, with symmetricKey: SymmetricKey?) throws
-}
-
-// MARK: - base types
-
-public protocol BaseContent: Codable {}
-extension Array : BaseContent where Element : BaseContent {}
-
-@available(iOS 13.0, *)
-public struct BaseMessage<C: BaseContent>: Codable {
-    public let uuid: UUID
-    public let sender: CoinbaseWalletSDK.PublicKey
-    public let content: C
-    public let version: String
 }
