+++ packages/wallet-native-sdk/ios/Message/Response/EncryptedResponseContent.swift
@@ -9,11 +9,11 @@ import Foundation
 import CryptoKit
 
 @available(iOS 13.0, *)
-public enum EncryptedResponseContent: EncryptedContent {
+enum EncryptedResponseContent: EncryptedContent {
     case response(requestId: UUID, data: Data)
     case failure(requestId: UUID, description: String)
     
-    public init(encrypt unencrypted: ResponseContent, with symmetricKey: SymmetricKey?) throws {
+    init(encrypt unencrypted: ResponseContent, with symmetricKey: SymmetricKey?) throws {
         switch unencrypted {
         case let .response(requestId, results):
             guard let symmetricKey = symmetricKey else {
@@ -29,7 +29,7 @@ public enum EncryptedResponseContent: EncryptedContent {
 
 @available(iOS 13.0, *)
 extension ResponseContent {
-    public init(decrypt encrypted: EncryptedResponseContent, with symmetricKey: SymmetricKey?) throws {
+    init(decrypt encrypted: EncryptedResponseContent, with symmetricKey: SymmetricKey?) throws {
         switch encrypted {
         case let .response(requestId, encryptedResults):
             guard let symmetricKey = symmetricKey else {
